+++ PluginInstallation/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [v1.4.0](https://github.com/packlink-dev/magento2_module/compare/v1.3.8...v1.4.0) - 2025-03-27
+### Added
+- Support for the special services
+- Manual refresh for shipping services
+### Changed
+- Updated to the latest Core version 3.5.0
+
 ## [1.3.7](https://github.com/packlink-dev/magento2_module/compare/v1.3.6...v1.3.8)
 ### Added
 - Compatibility with Magento 2.4.7